fetch wakatime on server
