+++ app/about/page.mdx
@@ -6,7 +6,7 @@ import Socials from "./socials";
 import U from "./underline";
 import WakaTime from "./wakatime";
 
-export const dynamic = "force-static";
+export const runtime = "edge";
 
 export const metadata = createMetadata({
   title: "About",
