+++ lib/schema.ts
@@ -9,14 +9,3 @@ export const PostsSchema = z.object({
   ),
 });
 export type Posts = z.infer<typeof PostsSchema>;
-
-export const WakaTimeLanguageSchema = z.object({
-  name: z.string(),
-  text: z.string(),
-  percent: z.number(),
-});
-
-export const WakaTimeSchema = z.object({
-  languages: z.array(WakaTimeLanguageSchema),
-});
-export type WakaTime = z.infer<typeof WakaTimeSchema>;
